Initial step to separate sync namespace refactoring and change namespace service
